Fix bug of columns

+++ includes/intersection.h
@@ -17,15 +17,7 @@
 # include "defines.h"
 
 t_byte	is_wall(t_wolf *wolf, unsigned short x, unsigned short y);
-t_point	intersection_horizontal(
-	t_wolf *wolf,
-	float angle,
-	unsigned short wall_y
-);
-t_point	intersection_vertical(
-	t_wolf *wolf,
-	float angle,
-	unsigned short wall_x
-);
+t_point	intersection_horizontal(t_wolf *wolf, float angle, float wall_y);
+t_point	intersection_vertical(t_wolf *wolf, float angle, float wall_x);
 
 #endif
